+++ RELEASE_NOTES.md
@@ -13,6 +13,7 @@
 * Added tag filtering to the `receive-gridpool-orders` command.
 * Added gridpool order ID filtering to `Client.list_gridpool_orders()`,
   `Client.gridpool_orders_stream()`, and the `receive-gridpool-orders` command.
+* Added staging service support to the `create-order` CLI command.
 
 ## Bug Fixes